moved soft eq/neq, documentation typos
