+++ chirho/observational/handlers/__init__.py
@@ -1,1 +1,2 @@
 from .condition import condition  # noqa: F401
+from .soft_conditioning import soft_eq, soft_neq  # noqa: F401
